refactor: move utxo functions to iotago.go
